requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
h5py==3.11.0
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.21.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
psutil==6.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.13.0
tbb==2021.13.0
threadpoolctl==3.5.0
torch==2.3.1
torch_geometric==2.5.3
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
zipp==3.19.2
